Another email verification db is added Break UsersService.kt into multiple services with register, verification and token management services The new Verification db stores the current user resend key by which the user can request another email send Removed the db as we need to made some changes for primary key to allow auto-increment TokenGenerator.kt helps to create a secure random number and also helper method to hash it. Most of the entity fields are made var

+++ src/main/resources/db/migrations/V1__include_initial_schema.sql
@@ -1,5 +1,5 @@
 CREATE TABLE users_table (
-    user_id BIGINT PRIMARY KEY,
+    user_id INTEGER PRIMARY KEY ,
     email VARCHAR(255) NOT NULL,
     p_word VARCHAR(255) NOT NULL,
     user_name VARCHAR(255) NOT NULL UNIQUE,
@@ -8,8 +8,8 @@ CREATE TABLE users_table (
 );
 
 CREATE TABLE users_profile (
-    profile_id BIGINT PRIMARY KEY,
-    user_id BIGINT NOT NULL UNIQUE,
+    profile_id INTEGER PRIMARY KEY ,
+    user_id INTEGER NOT NULL UNIQUE,
     bio VARCHAR(512),
     full_name VARCHAR(255),
     avatar_url VARCHAR(255),
